Reorganize utils and turn features into kwargs
