+++ tests/scripts/test_coarse_to_fine.py
@@ -6,7 +6,7 @@ import torch
 
 from fugw.scripts import coarse_to_fine
 from fugw.mappings import FUGW, FUGWSparse
-from fugw.mappings.utils import init_mock_distribution
+from fugw.utils import init_mock_distribution
 
 np.random.seed(0)
 torch.manual_seed(0)
